# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "小说创作应用后端接口文档。",
        "title": "小说创作应用 API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "使用配置文件中的系统密码登录，成功后返回 Bearer token。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handler_auth.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功",
                        "schema": {
                            "$ref": "#/definitions/auth.LoginSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/auth.ErrorBody"
                        }
                    },
                    "401": {
                        "description": "登录密码错误",
                        "schema": {
                            "$ref": "#/definitions/auth.ErrorBody"
                        }
                    }
                }
            }
        },
        "/novels": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "分页查询小说列表。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "novels"
                ],
                "summary": "查询小说列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "当前页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "$ref": "#/definitions/novel.ListSuccessResponse"
                        }
                    },
                    "401": {
                        "description": "未登录或登录已过期",
                        "schema": {
                            "$ref": "#/definitions/novel.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/novel.ErrorBody"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "创建一条小说记录。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "novels"
                ],
                "summary": "创建小说",
                "parameters": [
                    {
                        "description": "创建小说请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handler_novel.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/novel.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/novel.ErrorBody"
                        }
                    },
                    "401": {
                        "description": "未登录或登录已过期",
                        "schema": {
                            "$ref": "#/definitions/novel.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/novel.ErrorBody"
                        }
                    }
                }
            }
        },
        "/novels/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "根据 ID 查询小说详情。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "novels"
                ],
                "summary": "查询小说详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "小说 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "$ref": "#/definitions/novel.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/novel.ErrorBody"
                        }
                    },
                    "401": {
                        "description": "未登录或登录已过期",
                        "schema": {
                            "$ref": "#/definitions/novel.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "小说不存在",
                        "schema": {
                            "$ref": "#/definitions/novel.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/novel.ErrorBody"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "根据 ID 更新小说记录。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "novels"
                ],
                "summary": "更新小说",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "小说 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新小说请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handler_novel.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/novel.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/novel.ErrorBody"
                        }
                    },
                    "401": {
                        "description": "未登录或登录已过期",
                        "schema": {
                            "$ref": "#/definitions/novel.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "小说不存在",
                        "schema": {
                            "$ref": "#/definitions/novel.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/novel.ErrorBody"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "根据 ID 真删小说记录。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "novels"
                ],
                "summary": "删除小说",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "小说 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/novel.DeleteSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/novel.ErrorBody"
                        }
                    },
                    "401": {
                        "description": "未登录或登录已过期",
                        "schema": {
                            "$ref": "#/definitions/novel.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "小说不存在",
                        "schema": {
                            "$ref": "#/definitions/novel.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/novel.ErrorBody"
                        }
                    }
                }
            }
        },
        "/novels/{novel_id}/chapters": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "分页查询指定小说下的章节列表，列表项不包含正文。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chapters"
                ],
                "summary": "查询小说章节列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "小说 ID",
                        "name": "novel_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "当前页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "$ref": "#/definitions/chapter.ListSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/chapter.ErrorBody"
                        }
                    },
                    "401": {
                        "description": "未登录或登录已过期",
                        "schema": {
                            "$ref": "#/definitions/chapter.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "小说不存在",
                        "schema": {
                            "$ref": "#/definitions/chapter.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/chapter.ErrorBody"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "在指定小说下创建章节，章节号由后端自动分配。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chapters"
                ],
                "summary": "创建小说章节",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "小说 ID",
                        "name": "novel_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "创建章节请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handler_chapter.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/chapter.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/chapter.ErrorBody"
                        }
                    },
                    "401": {
                        "description": "未登录或登录已过期",
                        "schema": {
                            "$ref": "#/definitions/chapter.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "小说不存在",
                        "schema": {
                            "$ref": "#/definitions/chapter.ErrorBody"
                        }
                    },
                    "409": {
                        "description": "章节号冲突",
                        "schema": {
                            "$ref": "#/definitions/chapter.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/chapter.ErrorBody"
                        }
                    }
                }
            }
        },
        "/novels/{novel_id}/chapters/{chapter_id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "根据小说 ID 和章节 ID 查询章节详情，详情包含正文。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chapters"
                ],
                "summary": "查询小说章节详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "小说 ID",
                        "name": "novel_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "章节 ID",
                        "name": "chapter_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "$ref": "#/definitions/chapter.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/chapter.ErrorBody"
                        }
                    },
                    "401": {
                        "description": "未登录或登录已过期",
                        "schema": {
                            "$ref": "#/definitions/chapter.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "章节不存在",
                        "schema": {
                            "$ref": "#/definitions/chapter.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/chapter.ErrorBody"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "根据小说 ID 和章节 ID 更新章节标题和正文，字数由后端重新计算。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chapters"
                ],
                "summary": "更新小说章节",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "小说 ID",
                        "name": "novel_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "章节 ID",
                        "name": "chapter_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新章节请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handler_chapter.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/chapter.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/chapter.ErrorBody"
                        }
                    },
                    "401": {
                        "description": "未登录或登录已过期",
                        "schema": {
                            "$ref": "#/definitions/chapter.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "章节不存在",
                        "schema": {
                            "$ref": "#/definitions/chapter.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/chapter.ErrorBody"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "根据小说 ID 和章节 ID 删除章节，删除后不重排章节号。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chapters"
                ],
                "summary": "删除小说章节",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "小说 ID",
                        "name": "novel_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "章节 ID",
                        "name": "chapter_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/chapter.DeleteSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/chapter.ErrorBody"
                        }
                    },
                    "401": {
                        "description": "未登录或登录已过期",
                        "schema": {
                            "$ref": "#/definitions/chapter.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "章节不存在",
                        "schema": {
                            "$ref": "#/definitions/chapter.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/chapter.ErrorBody"
                        }
                    }
                }
            }
        },
        "/uploads/images": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "上传小说封面或人物图片到私有对象存储，并返回对象 key 和预签名预览链接。",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "uploads"
                ],
                "summary": "上传图片",
                "parameters": [
                    {
                        "type": "file",
                        "description": "图片文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "enum": [
                            "cover",
                            "character"
                        ],
                        "type": "string",
                        "description": "图片用途",
                        "name": "usage",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "上传成功",
                        "schema": {
                            "$ref": "#/definitions/upload.UploadImageSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/upload.ErrorBody"
                        }
                    },
                    "401": {
                        "description": "未登录或登录已过期",
                        "schema": {
                            "$ref": "#/definitions/upload.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/upload.ErrorBody"
                        }
                    }
                }
            }
        },
        "/uploads/preview": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "根据对象 key 生成新的私有图片预签名预览链接。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "uploads"
                ],
                "summary": "刷新图片预览链接",
                "parameters": [
                    {
                        "type": "string",
                        "description": "对象 key",
                        "name": "object_key",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "刷新成功",
                        "schema": {
                            "$ref": "#/definitions/upload.PreviewSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/upload.ErrorBody"
                        }
                    },
                    "401": {
                        "description": "未登录或登录已过期",
                        "schema": {
                            "$ref": "#/definitions/upload.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/upload.ErrorBody"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "auth.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code 表示错误响应码，使用 HTTP 状态码。",
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "description": "Message 表示用户可理解的错误提示。",
                    "type": "string",
                    "example": "请求参数错误"
                }
            }
        },
...
        "auth.LoginSuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code 表示业务响应码，成功固定为 0。",
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "description": "Data 表示登录成功后的令牌数据。",
                    "allOf": [
                        {
                            "$ref": "#/definitions/auth.LoginData"
                        }
                    ]
                },
                "message": {
                    "description": "Message 表示响应提示信息。",
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
        "internal_api_handler_auth.LoginRequest": {
            "type": "object",
            "required": [
                "password"
            ],
            "properties": {
                "password": {
                    "description": "Password 表示系统登录密码。",
                    "type": "string",
                    "example": "admin123"
                }
            }
        },
...
        "novel.ListSuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code 表示业务响应码，成功固定为 0。",
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "description": "Data 表示小说分页列表响应数据。",
                    "allOf": [
                        {
                            "$ref": "#/definitions/novel.ListData"
                        }
                    ]
                },
                "message": {
                    "description": "Message 表示响应提示信息。",
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "输入 Bearer token，例如：Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}